requirements.txt
pip==19.3.1
wheel==0.33.6
numpy==1.17.4
nibabel==3.0.0
pillow==7.0.0
twine=2.0.0
matplotlib==3.1.1